Consolidate events between EventConstants and BrowserEventEmitter
This is a follow up to
https://github.com/facebook/react/pull/9333. This commit removes some
duplication of event names and renames the EventConstants module to
BrowserEventConstants.
